+++ lib/src/components/NxDropdown/AbstractDropdown.tsx
@@ -4,13 +4,14 @@
  * the terms of the Eclipse Public License 2.0 which accompanies this
  * distribution and is available at https://www.eclipse.org/legal/epl-2.0/.
  */
-import React, { Children, forwardRef, useState } from 'react';
+import React, { Children, forwardRef } from 'react';
 import useMergedRef from '@react-hook/merged-ref';
 
 import { KeyboardEventHandler, useEffect, useRef } from 'react';
 import NxDropdownMenu from '../NxDropdownMenu/NxDropdownMenu';
 
 import { AbstractDropdownProps } from './types';
+import useDropdownOpenTop from '../../util/useDropdownOpenTop';
 export {
   AbstractDropdownProps,
   AbstractDropdownRenderToggleElement
@@ -133,34 +134,10 @@ const AbstractDropdown = forwardRef<HTMLDivElement, AbstractDropdownProps>((prop
     }
   }
 
-  useEffect(() => {
-    window.addEventListener('scroll', determineDropdownUp);
-    window.addEventListener('resize', determineDropdownUp);
-    return () => {
-      window.removeEventListener('scroll', determineDropdownUp);
-      window.removeEventListener('resize', determineDropdownUp);
-    };
-  }, []);
-
   const divRef = useRef<HTMLDivElement>(null);
   const newRef = useMergedRef(divRef, ref);
-  const [openTop, setOpenTop] = useState<boolean>(false);
-
-  const determineDropdownUp = () => {
-    if (divRef && divRef.current) {
-      const windowHeight = window.innerHeight;
-      const menuHeight = Children.count(children) * 32; //32px is the height of an item in the dropdown
-
-      const offset = divRef.current.getBoundingClientRect().bottom + menuHeight;
-
-      if (offset >= windowHeight) {
-        setOpenTop(true);
-      }
-      else {
-        setOpenTop(false);
-      }
-    }
-  };
+  const childrenCount = Children.count(children);
+  const openTop = useDropdownOpenTop(divRef, childrenCount);
 
   return (
     <div ref={newRef} className={className} onKeyDown={onKeyDown} {...attrs}>
